chore: Refactor screenshots script to extract func in files
