+++ .gitignore
@@ -25,4 +25,4 @@ transpiled
 # Generated files
 react/Icon/icons-sprite.js
 react/palette.js
-screenshots/
+/screenshots
